file-handle: Parse procinfo to get mount_fd
Right now I pass AT_FDCWD to open_by_handle_at(). This assumes that object
is reachable through same mountoint current working directory of process.
But this does not work if I open a file on a different file system (and
test program is running from a different fs). 

So parse /proc/self/procinfo and find the mount point and open it and get
an fd and pass that to open_by_handle_at().

Signed-off-by: Vivek Goyal <vgoyal@redhat.com>